Extract MailPoetLogoResponsive component
[MAILPOET-4421]
