+++ mailpoet/assets/js/src/common/top_bar/top_bar.tsx
@@ -1,10 +1,8 @@
 import { ReactNode } from 'react';
 import classnames from 'classnames';
-import { useHistory } from 'react-router-dom';
 import { t } from 'common/functions/t';
 import { withFeatureAnnouncement } from 'announcements/with_feature_announcement';
-import { MailPoetLogo } from './mailpoet_logo';
-import { MailPoetLogoMobile } from './mailpoet_logo_mobile';
+import { MailPoetLogoResponsive } from './mailpoet_logo_responsive';
 import { BeamerIcon } from './beamer_icon';
 import { ScreenOptionsFix } from './screen_options_fix';
 
@@ -19,33 +17,9 @@ export function TopBar({ children, hasNews, onBeamerClick }: Props) {
     'mailpoet-top-bar-beamer',
     hasNews ? 'mailpoet-top-bar-beamer-dot' : '',
   );
-  const history = useHistory();
-  const onLogoClick = () => history.push('/');
   return (
     <div className="mailpoet-top-bar">
-      <a
-        role="button"
-        className="mailpoet-top-bar-logo"
-        title={t('topBarLogoTitle')}
-        onClick={onLogoClick}
-        tabIndex={0}
-        onKeyDown={(event) => {
-          if (
-            ['keydown', 'keypress'].includes(event.type) &&
-            ['Enter', ' '].includes(event.key)
-          ) {
-            event.preventDefault();
-            onLogoClick();
-          }
-        }}
-      >
-        <div className="mailpoet-top-bar-logo-desktop">
-          <MailPoetLogo />
-        </div>
-        <div className="mailpoet-top-bar-logo-mobile">
-          <MailPoetLogoMobile />
-        </div>
-      </a>
+      <MailPoetLogoResponsive />
       <div className="mailpoet-top-bar-children">{children}</div>
       <div className="mailpoet-flex-grow" />
       {onBeamerClick && (
