+++ mailpoet/views/layout.html
@@ -127,6 +127,8 @@ jQuery('#adminmenu #toplevel_page_mailpoet-newsletters')
   'allSendingPausedBody': __('Your [link]API key[/link] to send with MailPoet is invalid.'),
   'allSendingPausedLink': __('Purchase a key'),
 
+  'topBarLogoTitle': __('Back to section root'),
+
   'close': __('Close'),
   'today': __('Today'),
   'january': __('January'),
