+++ mailpoet/views/newsletters.html
@@ -430,7 +430,6 @@
     'bounces': _x('Bounces', 'A tab title for the list of bounces (w.wiki/45Qc)'),
 
     'premiumBannerCtaFree': __('Upgrade'),
-    'topBarLogoTitle': __('Back to section root'),
 
     'gaCampaignLine': __('Google Analytics Campaign'),
     'gaCampaignTip': __('For example, “Spring email”. [link]Read the guide.[/link]'),
